+++ spawn/handler.py
@@ -13,8 +13,7 @@ from eigr.protocol_pb2 import Context
 import io
 import logging
 
-action_handler = Blueprint(
-    'action_endpoint', __name__, url_prefix='/api/v1')
+action_handler = Blueprint('action_endpoint', __name__, url_prefix='/api/v1')
 
 
 @action_handler.route('/system/<string:system>/actors/<string:name>/invoke', methods=["POST"])
