--- spawn/internal/__init__.py
@@ -1,4 +1,0 @@
-"""
-Copyright 2022 Eigr.
-Licensed under the Apache License, Version 2.0.
-"""
